+++ R/read_write.R
@@ -684,8 +684,8 @@ assign_to_global <- function(key, val, pos=1) {
 }
 
 
-#' @name save.ages
-#' @title save age summary of depth(s) 
+#' @name ageranges
+#' @title save age ranges of depth(s) 
 #' @description Calculates an age summary (min/max probability range, median and mean) of a depth or a series of depths. The depths are returned and can be saved to a file.  
 #' @return The minimum and maximum of the probability range (default 2.5% of both sides), median and mean ages for a depth or series of depths.
 #' @param d Depth or depths for which age summaries are to be calculated. If left empty, a sequence of depths is calculated from the top to the bottom of the core, see \link{d.by}.
@@ -693,16 +693,16 @@ assign_to_global <- function(key, val, pos=1) {
 #' @param sep Separator for the fields, if saving to a file (defaults to a tab, "\t").
 #' @param set Detailed information of the current run, stored within this session's memory as variable \code{info}.
 #' @param BCAD The calendar scale of graphs and age output-files is in cal BP (calendar or calibrated years before the present, where the present is AD 1950) by default, but can be changed to BC/AD using \code{BCAD=TRUE}.
-#' @param na.rm Whether or not NAs are to be removed. Defaults to \code{na.rm=FALSE}.
+#' @param na.rm Whether or not NAs are to be removed. Defaults to \code{na.rm=FALSE}. NAs will still be returned for depths outside of the core's depth range.
 #' @param prob Probability range. Half of the range is taken away from both sides of the distribution (e.g., 2.5\% for the default of \code{prob=0.95}).
 #' @param d.by Steps for calculation of depths, if \link{d} is left empty. Defaults to steps of 1. 
 #' @param roundby Rounding for the age estimates. Defaults to 1 decimal.
 #' @param show.progress Show a progress bar. Defaults to TRUE.
 #' @examples
-#'   myages <- save.ages(20:30)
+#'   myages <- ageranges(20:30)
 #'   myages
 #' @export
-save.ages <- function(d=c(), file=c(), sep="\t", set=get("info"), BCAD=set$BCAD, na.rm=FALSE, prob=0.95, d.by=1, roundby=1, show.progress=TRUE) {
+ageranges <- function(d=c(), file=c(), sep="\t", set=get("info"), BCAD=set$BCAD, na.rm=FALSE, prob=0.95, d.by=1, roundby=1, show.progress=TRUE) {
   if(length(d) == 0)
     d <- seq(set$d.min, set$d.max, by=d.by)
   
@@ -723,7 +723,7 @@ save.ages <- function(d=c(), file=c(), sep="\t", set=get("info"), BCAD=set$BCAD,
           setTxtProgressBar(pb, i)
 	  
     ages <- Bacon.Age.d(d[i], set, BCAD=BCAD, na.rm=na.rm)
-    quan <- quantile(ages, c((1-prob)/2, 1-((1-prob)/2), .5))
+    quan <- quantile(ages, c((1-prob)/2, 1-((1-prob)/2), .5), na.rm=na.rm)
     mins[i] <- quan[1]
     maxs[i] <- quan[2]
     medians[i] <- quan[3]
@@ -732,21 +732,22 @@ save.ages <- function(d=c(), file=c(), sep="\t", set=get("info"), BCAD=set$BCAD,
   if(show.progress && n>50)
     message("")
   
+  mins <- round(mins, roundby)
+  maxs <- round(maxs, roundby)
+  medians <- round(medians, roundby)
+  means <- round(means, roundby)
+  
   if(length(d) == 1) {
-    summ <- c(
-      min=round(mins, roundby),
-      max=round(maxs, roundby),
-      median=round(medians, roundby),
-      mean=round(means, roundby))
+    names <- c(paste0("min.", 100*prob), paste0("max.", 100 * prob),
+      "median", "mean")	  
+	summ <- setNames(c(mins, maxs, medians, means), names)
   } else {
-      summ <- data.frame(
-        depth=round(d, roundby),
-        min=round(mins, roundby),
-        max=round(maxs, roundby),
-        median=round(medians, roundby),
-        mean=round(means, roundby))
-      }
+      summ <- data.frame(depth = d, mins, maxs, medians, means)
+      colnames(summ) <- c("depth",
+        paste0("min.", 100 * prob), paste0("max.", 100*prob),
+        "median", "mean")
+    }
   if(length(file) > 0)
-    write.table(summ, file, row.names=FALSE, sep=sep, quote=FALSE)
+    fastwrite(summ, file, col.names=TRUE, row.names=FALSE, sep=sep, quote=FALSE)
   return(summ)	  
 }